+++ server/src/api/search/routes.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from api.common.dependencies import validated_token, SpotifyClient
+from api.common.dependencies import validated_token
 from api.search.dependencies import SearchSpotifyClient
 from api.search.models import GeneralSearchResult, PaginatedSearchResult, Playlist, Artist, Album, Track, \
     SpotifyPlayableType
@@ -14,24 +14,28 @@ router = APIRouter(
 @router.get("/")
 async def search(query: str, token: validated_token, spotify_client: SearchSpotifyClient) -> GeneralSearchResult:
     all_playable_types = [e.value for e in SpotifyPlayableType]
-    return spotify_client.get_search(query, token, all_playable_types)
+    return spotify_client.get_general_search(query, token, all_playable_types)
 
 
 @router.get("/tracks")
-async def search_tracks(query: str, limit: int = 20, offset: int = 0) -> PaginatedSearchResult[Track]:
-    pass
+async def search_tracks(token: validated_token, spotify_client: SearchSpotifyClient, query: str, offset: int = 0,
+                        limit: int = 20) -> PaginatedSearchResult[Track]:
+    return spotify_client.get_track_search(query, token, offset, limit)
 
 
 @router.get("/albums")
-async def search_albums(query: str, limit: int = 20, offset: int = 0) -> PaginatedSearchResult[Album]:
+async def search_albums(token: validated_token, query: str, limit: int = 20, offset: int = 0) \
+        -> PaginatedSearchResult[Album]:
     pass
 
 
 @router.get("/artists")
-async def search_artists(query: str, limit: int = 20, offset: int = 0) -> PaginatedSearchResult[Artist]:
+async def search_artists(token: validated_token, query: str, limit: int = 20, offset: int = 0) \
+        -> PaginatedSearchResult[Artist]:
     pass
 
 
 @router.get("/playlists")
-async def search_playlists(query: str, limit: int = 20, offset: int = 0) -> PaginatedSearchResult[Playlist]:
+async def search_playlists(token: validated_token, query: str, limit: int = 20, offset: int = 0) \
+        -> PaginatedSearchResult[Playlist]:
     pass
